Breaking: belatedly add means for handling of the baseUrl changes in 8
The changes in version 8 of the generator (necessary to allow it to
be functional against recent versions of metalsmith and its plugins)
made the output base all paths at '/'. It turns out that some plugins
(notably u-dom) publish their documents into a subfolder of the main
unexpected documentation site. Support this use-case via a command
line --base-url option which can also be a path like '/plugin'.